Allow setting of azimuth/elevation angles
This can be used to display the azimuth pattern at a different elevation
angle and/or the elevation pattern at a different azimuth angle than the
default at maximum.

Bug-fix in computation of maximum gain azimuth direction: If the maximum
gain in theta direction goes up or down, the azimuth angle would be
computed incorrectly because all gain values at that theta angle are the
same for all azimuth angles.