Use ppm images for testing
The png image from matplotlib contains the matplotlib version.
The ppm format (converted from the png using PIL) doesn't have the
matplotlib version and therefore we're more likely to have a successful
test by fuzzy matching of versions. If the version of matplotlib is not
exactly known we interpolate the version and return up to two matching
candidates. This still won't ensure we always do the correct test with
the correct matplotlib version but it improves things.